ir+metainterp: move PreambleOp + FieldEntry to majit-ir
Pure-data types stored inside `PtrInfo` struct/array field caches and
the short-preamble Phase 2 sentinel. Only references `Op` and `OpRef`,
both already in majit-ir. info.rs re-exports the names so consumer
imports stay valid.

Prepares the PtrInfo move that follows — `InstancePtrInfo.fields:
Vec<(u32, FieldEntry)>` and the parallel slots on Struct/Array need
`FieldEntry` to be majit-ir-local.
